backend/requirements.txt
PyMuPDF
google-cloud-aiplatform
google-cloud-texttospeech
pydub
flask
python-dotenv